frontend/requirements.txt
pandas
numpy
regex
yake
streamlit
Pillow
requests
replicate
python-dotenv
